requirements.txt
asgiref==3.3.4
blis==0.7.4
catalogue==2.0.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cymem==2.0.5
Django==3.0.5
django-jazzmin==2.4.6
djongo==1.3.4
dnspython==2.1.0
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.0.0/en_core_web_md-3.0.0-py3-none-any.whl
idna==2.10
Jinja2==3.0.0
MarkupSafe==2.0.0
murmurhash
numpy
opencv-python
packaging==20.9
pandas==1.2.4
pathy==0.5.2
Pillow==8.2.0
preshed
pydantic==1.7.4
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
smart-open
spacy
spacy-legacy==3.0.5
sqlparse==0.2.4
srsly
thinc==8.0.3
tqdm==4.60.0
typer==0.3.2
urllib3==1.26.4
wasabi
